Includes Co, Nb, and Mo data
Thermodynamic data for the metals Co, Nb, and Mo as well as their fluorides are included.
The mapped std::array in the std::unordered_map heatData increases in size (from 8 to 9), since the heat capacity of two species includes a 1/T term, which was not present in any of the pre-existing species. The std::array size parameter is now stored in a const variable.
Associated functions with the new metals are added.