Changes to the decoupleSurr function
The decoupleSurr function now calculates the fluorides of Co and Mo since their thermodynamic data are now available. Work on Nb will be included soon.
Equilibrium including solids from now on will be calculated such that all solid mole fractions are affected by the leaching of metal into the salt and gas phases and are therefore not treated as constants.